[#RSL] Accurate casting

+++ src/headers/Lock.h
@@ -14,13 +14,13 @@ namespace TDA
         std::string m_session_token;
         double m_lifeTime;
         std::chrono::_V2::system_clock::time_point m_start;
-        
+
     private:
         std::string createToken();
 
     public:
         Lock(std::string _apiKey, std::string _name, double _lifeTime);
-        Lock(uint32_t _id, std::string _apiKey, std::string _name, std::string _session_token, double _lifeTime);
+        Lock(long unsigned int _id, std::string _apiKey, std::string _name, std::string _session_token, double _lifeTime);
         //~Lock();
 
         bool expired();
